k8s/talos/requirements.txt
jsonpatch==1.32
loguru>=0.5.3
pyyaml>=6.0
typer>=0.4.0
